requirements.txt
# watsonx.ai Learning Hub - Python Dependencies
# Houston Community College

# Core watsonx.ai SDK
ibm-watsonx-ai>=1.0.0

# IBM Cloud SDK
ibm-cloud-sdk-core>=3.16.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Jupyter and Notebook Support
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
ipywidgets>=8.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning Frameworks
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Deep Learning (Optional - install if needed)
# tensorflow>=2.13.0
# torch>=2.0.0
# keras>=2.13.0

# LangChain Integration
langchain>=0.1.0
langchain-community>=0.0.20
langchain-ibm>=0.1.0

# LlamaIndex Integration
llama-index>=0.9.0
llama-index-llms-ibm>=0.1.0

# Vector Stores and Embeddings
faiss-cpu>=1.7.4
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Document Processing
pypdf>=3.15.0
python-docx>=0.8.11
python-pptx>=0.6.21

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0
pyyaml>=6.0

# Testing (Optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0

# Code Quality (Optional)
# black>=23.7.0
# flake8>=6.1.0
# mypy>=1.5.0